microservices/storage_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
sqlalchemy